+++ src/main/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3/SearchCriteria.java
@@ -1,0 +1,8 @@
+package student_aleksandrs_korsaks.ak_lesson_10.level_2_3;
+
+interface SearchCriteria {
+
+    boolean match(Book book);
+
+
+}
